abstracted out Solr components into a CrawlDbProxyFactory
