+++ sparkler-core/sparkler-app/src/main/scala/edu/usc/irds/sparkler/CrawlDbRDD.scala
@@ -55,7 +55,7 @@ class CrawlDbRDD(sc: SparkContext,
     query.set("sort", sortBy)
     query.setRows(batchSize)
 
-    new SolrResultIterator[Resource](job.newCrawlDbSolrClient().crawlDb, query,
+    new SolrResultIterator[Resource](job.newCrawlDbProxy().getClient(), query,
       batchSize, classOf[Resource], closeClient = true, limit = topN)
   }
 
@@ -68,8 +68,8 @@ class CrawlDbRDD(sc: SparkContext,
     qry.set("group.field", groupBy)
     qry.set("group.limit", 0)
     qry.setRows(maxGroups)
-    val proxy = job.newCrawlDbSolrClient()
-    val solr = proxy.crawlDb
+    val proxy = job.newCrawlDbProxy()
+    val solr = proxy.getClient()
     val groupRes = solr.query(qry).getGroupResponse.getValues.get(0)
     val grps = groupRes.getValues
     CrawlDbRDD.LOG.info(s"selecting ${grps.size()} out of ${groupRes.getNGroups}")
