+++ sparkler-core/sparkler-app/src/main/scala/edu/usc/irds/sparkler/pipeline/Crawler.scala
@@ -24,9 +24,8 @@ import edu.usc.irds.sparkler._
 import edu.usc.irds.sparkler.base.{CliTool, Loggable}
 import edu.usc.irds.sparkler.model.ResourceStatus._
 import edu.usc.irds.sparkler.model.{CrawlData, Resource, ResourceStatus, SparklerJob}
-import edu.usc.irds.sparkler.service.SolrProxy
 
-import edu.usc.irds.sparkler.storage.solr.{SolrStatusUpdate, SolrUpsert, StatusUpdateSolrTransformer, ScoreUpdateSolrTransformer}
+import edu.usc.irds.sparkler.storage.solr.{SolrProxy, SolrStatusUpdate, SolrUpsert, StatusUpdateSolrTransformer, ScoreUpdateSolrTransformer}
 import edu.usc.irds.sparkler.util.{JobUtil, NutchBridge}
 import org.apache.hadoop.conf.Configuration
 import org.apache.hadoop.io.Text
@@ -171,7 +170,7 @@ class Crawler extends CliTool {
     //STEP : Initialize environment
     init()
 
-    val solrc = this.job.newCrawlDbSolrClient()
+    val solrc = this.job.newCrawlDbProxy()
     LOG.info("Committing crawldb..")
     solrc.commitCrawlDb()
     val localFetchDelay = fetchDelay
