+++ sparkler-core/sparkler-app/src/main/scala/edu/usc/irds/sparkler/storage/solr/SolrUpsert.scala
@@ -35,7 +35,7 @@ class SolrUpsert(job: SparklerJob) extends ((TaskContext, Iterator[Resource]) =>
 
   override def apply(context: TaskContext, docs: Iterator[Resource]): Any = {
     LOG.debug("Inserting new resources into CrawlDb")
-    val solrClient = job.newCrawlDbSolrClient()
+    val solrClient = job.newCrawlDbProxy()
 
     //TODO: handle this in server side - tell solr to skip docs if they already exist
 
@@ -43,7 +43,7 @@ class SolrUpsert(job: SparklerJob) extends ((TaskContext, Iterator[Resource]) =>
     val newLinksFilter: (Resource => Boolean) = doc => {
       val qry = new SolrQuery(s"${Constants.solr.DEDUPE_ID}:${doc.getDedupeId}")
       qry.setRows(0) //we are interested in counts only and not the contents
-      solrClient.crawlDb.query(qry).getResults.getNumFound == 0
+      solrClient.getClient().query(qry).getResults.getNumFound == 0
       // if zero hits, then there are no duplicates
     }
     val newResources = docs.withFilter(newLinksFilter)
